+++ restart-server.sh
@@ -1,0 +1,9 @@
+#!/bin/bash
+
+echo Restarting servers...
+
+echo "---- Stopping ----"
+./stop-server.sh
+echo " "
+echo "---- Starting ----"
+./start-server.sh