+++ CHANGELOG.md
@@ -2,5 +2,8 @@
 
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+- Implement web server in Flask
+
 ## [1.0.0] - 2017-02-10
 - Initial release